[MIRROR] Clockies now can summon shields to bash nonbelievers into orbit. (#2033)
* Clockies now can summon shields to bash nonbelievers into orbit. (#12252)

* Shield time

Adds a shield for clockies which empowers bashes via the damage it blocks. Might or might not compile yet. Icons going to be added in the next commit because I'd rather not have to revert all instead of just the icons if I do a fucky

* Icon time

Adds the icons for the clockie shield. Just the implementation + icons as of now, no way to get it. Might still not compile :flushed:

* Compilation issue resolution

Now it compiles. So it works. Maybe.

* Little change

Only harm-intent bashes amplify the bash / use energy.

* Fixes up shield.dms repeair proc and moves the brass shield repair there

Oh cool time to throw in a fix aswell, neat!

* Some tweaks

Buffs the shields armor and integrity since it was VERY easy to break with energy before, now its medium against those and very strong against ballistics. also a bit easier to accumulate charge because its ''only'' 50% block

* Some more stuff for the shield

Adds a spell to summon the shield (20dc time, 600kw cost, tier-two spell)
aswell as adding some ''fun'' for non-clockie users of the shield.

* typo man bad

* more spellchecking

* even more little tweaks

* clarification on a comment

* sound tweak

only plays a special sound if its an empowered bash

* final tweakies before PR

* Abuse safeguard

Yeeah might be best to not allow clockies to bash eachother for infinite charge.

* another typo fix

I really shouldn't have been coding stuff for this at 5 am...

* Safeguards for a VERY specific edge case

Would have caused some weird behavious for edge cases if not resetting the values, so I'm doing it to be sure.

* 5am code bad

same thing like the previous commit but not unreachable because of return statements.

* Applying requested changes / suggestions proposed by Ghommie

Early return statement done by the harmful / servant check, other returns / simillar done via . = ..()

* Formatting

* Clockies now can summon shields to bash nonbelievers into orbit.

Co-authored-by: DeltaFire15 <46569814+DeltaFire15@users.noreply.github.com>