Changes to Shark Light snout
Makes the snout more snubbier, like originally intended.
This sprite was originaly mine, so I'm not pulling "like originally intended" out of nowhere, my intention was to make a snout that's visible, but snubby and not too pronounced.
The issue with the sprite was the eyes were too far back, the nose bridge was too far forward.