Fixed button, clarified text, cleaned code
"Move it to the threshold" button now does what it says. Removed any mention of "exponential slowdown" since it is not and never was exponential. Reworded things to be more clear to the player. Fixed some nested code that shouldn't be nested and added a return for the "no" option instead of letting the conditional run to a dead end. (I'm not sure if this is good or bad practice.)