Added Skyrat comments. Changed some code.
Removed unneeded return. Added a check for an empty `dorfy` variable which `return`s/aborts if true, which is more desirable than the original PR's method of "move to threshold" for this scenario. Also added a dumb fix that (hopefully) stops the below-threshold warning from displaying when you set your size to the exact threshold value.