Polish
Placed an invisible light in the middle of the killzone so it doesn't have a dark spot in the middle.

Replaced the airhorn with some hotdogs.

Made the contents of the tables adjacent to the safezone tiles a little less memey. They now have a delicious selection of donuts as opposed to toys and salt shakers.