Refurbished Centcom Docking Area - The Squeakquel (#3220)
* Initializing

Haven't used the map merger yet. StrongDMM is /supposed/ to do that automatically.

Moved less shit around this time.

* Changelog

:astonished:

* Replaces broken corndogs

As hilarious as it is for the delicious phallus shaped snacks to poof into thin air when you try to eat them, I want this PR to be as snag free as possible.

* Deleting Changelog

Apparently I don't need to do this anymore and nobody told me.

* More lights, detailing, and

Sometimes the lights spawn broken and that results in dark spots. Added more light tubes to compensate.

Also added some chairs, a water cooler, and some cool detailing to the neutral hallway to make it feel less bland.

Also adding a bunch of extinguisher cabinets seemed like a good idea too, considering how often people in the safezone end up on fire.

* Polish

Placed an invisible light in the middle of the killzone so it doesn't have a dark spot in the middle.

Replaced the airhorn with some hotdogs.

Made the contents of the tables adjacent to the safezone tiles a little less memey. They now have a delicious selection of donuts as opposed to toys and salt shakers.