Housekeeping for the weapons crafting menu
Declutters the CAT_WEAPON category, which had previously been a catch-all before the subcategories were added. This should put most if not all of the leftovers of the switch to the subcategory system into their proper subcategories. 

Also, see the comment on the craftable switchblade. Its recipe hasn't been updated since the guncrafting update. There's a more appropriate part for the recipe now but it might dick around with game balance because that part's cheaper to make. I'll leave that change up to someone who's been contributing longer than I have.