Revert "Added midround incursion sloppily"
This reverts commit dbcd79cd6591a747efa2006a737c3136d75de66f.

It was a bad idea, people not wanting antag as they are in the middle of other plans is a pretty common issue.
And having it be a group thing makes it worse to deal with. because if one is not wanting it, then nobody can be a incursionist due to meta-gaming and such, along with the fact there might be a few bugs with the uplink not spawning and bothering the staff with it every round is a bad thing.