Boxstation Security - New Prison & Aesthetic Overhaul (#3558)
* Box Prison Rework

Reworks prison and some of security

* Missing light

* Missing lights & misc bugfixes

* Shutters to the pool

Sec and prisoners never use firelocks to stop all the air escaping, annoyingly. This should help.

* Protolathe moved to armoury

* Fills up the tables

O2 tanks and metal/glass for the autolathe.

* Pool Controller moved

It's now on the other side of the pool so you don't have to swim through to get to it.